# backend/src/core/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Phong Thủy & Gieo Quẻ API",
    "version": "2.0.0",
    "description": "Tài liệu API và môi trường thử nghiệm trực quan cho Hệ thống Phong Thủy & Gieo Quẻ. Hỗ trợ đầy đủ các phân hệ Kinh Dịch Lục Hào/Mai Hoa, Tứ Trụ Bát Tự, Mệnh Bàn Tử Vi, Bát Tự Hợp Hôn, Xem Ngày Hoàng Đạo (Trạch Cát), Quản lý Thư mục (Tags), Bài viết (Blog), Luận giải AI Realtime (SSE) và Quản Trị Hệ Thống (Admin)."
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Máy chủ Phát triển Local"
    },
    {
      "url": "https://tuynover.ddns.net",
      "description": "Máy chủ Production (AWS EC2)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Nhập JWT token dạng: `Authorization: Bearer <token>`"
      },
      "queryAuth": {
        "type": "apiKey",
        "in": "query",
        "name": "token",
        "description": "JWT token truyền trực tiếp dưới dạng tham số truy vấn: `?token=<token>` (sử dụng cho luồng Server-Sent Events - SSE)"
      }
    },
    "schemas": {
...
      "ZiweiChartRequest": {
        "type": "object",
        "required": [
          "date",
          "hour",
          "gender",
          "calendarType"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "Nguyen Van A"
          },
          "date": {
            "type": "string",
            "example": "2004-09-05"
          },
          "hour": {
            "type": "integer",
            "minimum": 0,
            "maximum": 11,
            "example": 7
          },
          "gender": {
            "type": "string",
            "enum": [
              "Nam",
              "Nữ"
            ],
            "example": "Nam"
          },
          "calendarType": {
            "type": "string",
            "enum": [
              "solar",
              "lunar"
            ],
            "example": "solar"
          },
          "userId": {
            "type": "string",
            "example": "guest"
          }
        }
      },
...
      "CreateBlogPostRequest": {
        "type": "object",
        "required": [
          "title",
          "summary",
          "content"
        ],
        "properties": {
          "title": {
            "type": "string",
            "example": "Giải mã ý nghĩa 64 quẻ Kinh Dịch"
          },
          "slug": {
            "type": "string",
            "example": "giai-ma-y-nghia-64-que-kinh-dich"
          },
          "summary": {
            "type": "string",
            "example": "Tóm lược các nguyên lý căn bản của Kinh Dịch trong dự đoán sự nghiệp và tài vận."
          },
          "content": {
            "type": "string",
            "example": "# Nội dung bài viết Markdown..."
          },
          "category": {
            "type": "string",
            "example": "Kinh Dịch"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "Kinh Dịch",
              "Dự đoán"
            ]
          },
          "thumbnailUrl": {
            "type": "string",
            "example": "https://example.com/thumb.jpg"
          },
          "isPublished": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "UpdateBlogPostRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Giải mã chi tiết 64 quẻ Kinh Dịch"
          },
          "slug": {
            "type": "string",
            "example": "giai-ma-chi-tiet-64-que-kinh-dich"
          },
          "summary": {
            "type": "string",
            "example": "Tóm lược cập nhật..."
          },
          "content": {
            "type": "string",
            "example": "# Nội dung sửa đổi..."
          },
          "category": {
            "type": "string",
            "example": "Kinh Dịch"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "Kinh Dịch"
            ]
          },
          "thumbnailUrl": {
            "type": "string",
            "example": "https://example.com/thumb2.jpg"
          },
          "isPublished": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/ziwei": {
      "post": {
        "tags": [
          "Tính toán & Gieo Quẻ"
        ],
        "summary": "Lập Mệnh Bàn Tử Vi Đẩu Số (12 Cung & An Sao)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ZiweiChartRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Thành công. Trả về cấu trúc 12 cung, chính tinh, phụ tinh, tứ hóa và đại vận."
          }
        }
      }
    },
...
    "/api/blog": {
      "get": {
        "tags": [
          "Bài Viết & Kiến Thức (Blog)"
        ],
        "summary": "Lấy danh sách bài viết kiến thức phong thủy (Hỗ trợ phân trang, lọc chủ đề, tìm kiếm)",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 9
            }
          },
          {
            "name": "showAll",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Danh sách bài viết kèm metadata phân trang"
          }
        }
      },
      "post": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Bài Viết & Kiến Thức (Blog)"
        ],
        "summary": "Tạo bài viết kiến thức mới (Dành cho Admin/Co-Admin, tự động ping Googlebot)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBlogPostRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tạo bài viết thành công"
          },
          "403": {
            "description": "Không có quyền Admin"
          }
        }
      }
    },
...
    "/api/blog/{slug}": {
      "get": {
        "tags": [
          "Bài Viết & Kiến Thức (Blog)"
        ],
        "summary": "Lấy chi tiết bài viết theo Slug đường dẫn thân thiện (Tự động tăng lượt xem nguyên tử)",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "giai-ma-y-nghia-64-que-kinh-dich"
          }
        ],
        "responses": {
          "200": {
            "description": "Nội dung bài viết kèm danh sách bài liên quan"
          },
          "404": {
            "description": "Không tìm thấy bài viết"
          }
        }
      }
    },
    "/api/blog/{id}": {
      "put": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Bài Viết & Kiến Thức (Blog)"
        ],
        "summary": "Cập nhật bài viết theo ID (Admin only)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBlogPostRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cập nhật bài viết thành công"
          }
        }
      },
      "delete": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Bài Viết & Kiến Thức (Blog)"
        ],
        "summary": "Xóa mềm bài viết (Admin only, tự động ping URL_DELETED tới Google)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Xóa bài viết thành công"
          }
        }
      }
    },
...
  }
}